+++ kutana/context.py
@@ -12,10 +12,6 @@ class Context:
     :ivar ~.backend: Backend who created update.
     :ivar ~.config: Application's configuration
     :ivar ~.default_target_id: Default receiver of responses
-    :ivar ~.group_uid: Group's id plus backend identity
-    :ivar ~.user_uid: User's id plus backend identity
-    :ivar ~.group_state: State of group targeted by update
-    :ivar ~.user_state: State of user targeted by update
     """
 
     def __init__(self, app=None, config=None, update=None, backend=None):
@@ -25,11 +21,6 @@ class Context:
         self.config = config
 
         self.default_target_id = None
-        self.group_uid = None
-        self.user_uid = None
-
-        self.group_state = None
-        self.user_state = None
 
     @staticmethod
     async def create(app=None, config=None, update=None, backend=None):
@@ -43,31 +34,10 @@ class Context:
         )
 
         if update.type == ut.MSG:
-            backend_identity = backend.get_identity()
-            storage = app.storage
-
             if update.receiver_type == rt.MULTI:
                 ctx.default_target_id = update.receiver_id
-
-                ctx.group_uid = f"{update.receiver_id}{backend_identity}"
-                ctx.user_uid = f"{update.receiver_id}{backend_identity}"
-
-                ctx.group_state_key = f"_sg:{ctx.group_uid}"
-                ctx.user_state_key = f"_su:{ctx.user_uid}"
-
-                ctx.group_state = await storage.load(ctx.group_state_key, "")
-                ctx.user_state = await storage.load(ctx.user_state_key, "")
             else:
                 ctx.default_target_id = update.sender_id
-
-                ctx.group_uid = ""
-                ctx.user_uid = f"{update.sender_id}{backend_identity}"
-
-                ctx.group_state_key = ""
-                ctx.user_state_key = f"_su:{ctx.user_uid}"
-
-                ctx.group_state = ""
-                ctx.user_state = await storage.load(ctx.user_state_key, "")
 
         return ctx
 
@@ -172,24 +142,3 @@ class Context:
         ))
 
         return responses
-
-    async def set_state(self, group_state=None, user_state=None):
-        """
-        Set group_state and/or user_state for current group and/or user.
-        States can be represented by any string or None (if it shouldn't
-        change).
-        """
-
-        if group_state is not None:
-            if not isinstance(group_state, str):
-                raise ValueError("'group_state' must be string")
-            if self.group_uid == "":
-                raise ValueError("You can't set group state in solo chat")
-            await self.app.storage.save(self.group_state_key, group_state)
-
-        if user_state is not None:
-            if not isinstance(user_state, str):
-                raise ValueError("'user_state' must be string")
-            if self.user_uid == "":
-                raise ValueError("You can't set group state in solo chat")
-            await self.app.storage.save(self.user_state_key, user_state)