+++ kutana/storages/__init__.py
@@ -1,3 +1,0 @@
-from .naive_memory import NaiveMemory
-
-__all__ = ["NaiveMemory"]