+++ tests/test_router.py
@@ -5,21 +5,6 @@ from kutana.router import Router, ListRouter, MapRouter
 from kutana.routers import AnyMessageRouter
 
 
-def test_check():
-    r = Router()
-    h = Handler(None, "", "", 0)
-
-    c = Context()
-    c.group_state = ""
-    c.user_state = ""
-
-    assert r._check(h, None, c)
-
-    c.group_state = ""
-    c.user_state = "state"
-    assert not r._check(h, None, c)
-
-
 def test_exception_on_wrong_merge():
     lr = ListRouter()
     mr = MapRouter()
@@ -34,9 +19,9 @@ def test_exception_on_wrong_merge():
 def test_multiple_handlers():
     mr = MapRouter()
 
-    mr.add_handler(Handler(1, "*", "*", 0), "a")
-    mr.add_handler(Handler(2, "*", "*", 0), "a")
-    mr.add_handler(Handler(3, "*", "*", 0), "a")
+    mr.add_handler(Handler(1, 0), "a")
+    mr.add_handler(Handler(2, 0), "a")
+    mr.add_handler(Handler(3, 0), "a")
 
     assert len(mr._handlers["a"]) == 3
 
@@ -45,9 +30,9 @@ def test_router_merge():
     mr1 = MapRouter()
     mr2 = MapRouter()
 
-    mr1.add_handler(Handler(1, "*", "*", 0), "a")
-    mr1.add_handler(Handler(2, "*", "*", 0), "b")
-    mr2.add_handler(Handler(3, "*", "*", 0), "a")
+    mr1.add_handler(Handler(1, 0), "a")
+    mr1.add_handler(Handler(2, 0), "b")
+    mr2.add_handler(Handler(3, 0), "a")
 
     mr1.merge(mr2)
 
@@ -64,11 +49,11 @@ def test_router_subclass_merge():
     mr3 = ListRouter()
     mr4 = AnyMessageRouter()
 
-    mr1.add_handler(Handler(1, "*", "*", 0))
-    mr1.add_handler(Handler(2, "*", "*", 0))
-    mr2.add_handler(Handler(3, "*", "*", 0))
-    mr3.add_handler(Handler(4, "*", "*", 0))
-    mr4.add_handler(Handler(5, "*", "*", 0))
+    mr1.add_handler(Handler(1, 0))
+    mr1.add_handler(Handler(2, 0))
+    mr2.add_handler(Handler(3, 0))
+    mr3.add_handler(Handler(4, 0))
+    mr4.add_handler(Handler(5, 0))
 
     with pytest.raises(RuntimeError):
         mr1.merge(mr2)
@@ -94,12 +79,12 @@ def test_router_merge_priority():
     mr3 = AnyMessageRouter(priority=5)
     mr4 = AnyMessageRouter(priority=5)
 
-    mr1.add_handler(Handler(1, "*", "*", 0))
-    mr2.add_handler(Handler(2, "*", "*", 0))
-    mr3.add_handler(Handler(3, "*", "*", 0))
-    mr3.add_handler(Handler(4, "*", "*", 5))
-    mr4.add_handler(Handler(5, "*", "*", 3))
-    mr4.add_handler(Handler(6, "*", "*", 10))
+    mr1.add_handler(Handler(1, 0))
+    mr2.add_handler(Handler(2, 0))
+    mr3.add_handler(Handler(3, 0))
+    mr3.add_handler(Handler(4, 5))
+    mr4.add_handler(Handler(5, 3))
+    mr4.add_handler(Handler(6, 10))
 
     with pytest.raises(RuntimeError):
         mr1.merge(mr2)